Move app structs with nesting into API alongside existing non-nested structs
