+++ thoth-api/src/publication/model.rs
@@ -6,10 +6,12 @@ use strum::EnumString;
 use uuid::Uuid;
 
 use crate::graphql::utils::Direction;
+use crate::price::model::Price;
 #[cfg(feature = "backend")]
 use crate::schema::publication;
 #[cfg(feature = "backend")]
 use crate::schema::publication_history;
+use crate::work::model::WorkExtended as Work;
 
 #[cfg_attr(feature = "backend", derive(DbEnum, juniper::GraphQLEnum))]
 #[cfg_attr(feature = "backend", DieselType = "Publication_type")]
@@ -67,6 +69,48 @@ pub struct Publication {
     pub publication_url: Option<String>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct DetailedPublication {
+    pub publication_id: Uuid,
+    pub publication_type: PublicationType,
+    pub work_id: Uuid,
+    pub isbn: Option<String>,
+    pub publication_url: Option<String>,
+    pub updated_at: DateTime<Utc>,
+    pub work: Work,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct PublicationExtended {
+    pub publication_id: Uuid,
+    pub publication_type: PublicationType,
+    pub work_id: Uuid,
+    pub isbn: Option<String>,
+    pub publication_url: Option<String>,
+    pub prices: Option<Vec<Price>>,
+    pub work: SlimWork,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimWork {
+    pub imprint: SlimImprint,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimImprint {
+    pub publisher: SlimPublisher,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimPublisher {
+    pub publisher_id: Uuid,
 }
 
 #[cfg_attr(
@@ -135,6 +179,34 @@ impl Default for PublicationType {
 impl Default for PublicationField {
     fn default() -> Self {
         PublicationField::PublicationType
+    }
+}
+
+impl Default for DetailedPublication {
+    fn default() -> DetailedPublication {
+        DetailedPublication {
+            publication_id: Default::default(),
+            publication_type: Default::default(),
+            work_id: Default::default(),
+            isbn: None,
+            publication_url: None,
+            updated_at: chrono::TimeZone::timestamp(&Utc, 0, 0),
+            work: Default::default(),
+        }
+    }
+}
+
+impl Default for PublicationExtended {
+    fn default() -> PublicationExtended {
+        PublicationExtended {
+            publication_id: Default::default(),
+            publication_type: PublicationType::Paperback,
+            work_id: Default::default(),
+            isbn: None,
+            publication_url: None,
+            prices: Default::default(),
+            work: Default::default(),
+        }
     }
 }
