+++ thoth-app/src/component/issues_form.rs
@@ -1,4 +1,6 @@
 use thoth_api::account::model::AccountDetails;
+use thoth_api::issue::model::IssueExtended as Issue;
+use thoth_api::series::model::SeriesExtended as Series;
 use uuid::Uuid;
 use yew::html;
 use yew::prelude::*;
@@ -24,13 +26,11 @@ use crate::models::issue::delete_issue_mutation::DeleteIssueRequestBody;
 use crate::models::issue::delete_issue_mutation::PushActionDeleteIssue;
 use crate::models::issue::delete_issue_mutation::PushDeleteIssue;
 use crate::models::issue::delete_issue_mutation::Variables as DeleteVariables;
-use crate::models::issue::Issue;
 use crate::models::series::serieses_query::FetchActionSerieses;
 use crate::models::series::serieses_query::FetchSerieses;
 use crate::models::series::serieses_query::SeriesesRequest;
 use crate::models::series::serieses_query::SeriesesRequestBody;
 use crate::models::series::serieses_query::Variables;
-use crate::models::series::Series;
 use crate::models::Dropdown;
 use crate::string::CANCEL_BUTTON;
 use crate::string::EMPTY_ISSUES;
