+++ thoth-app/src/component/new_series.rs
@@ -1,5 +1,7 @@
 use std::str::FromStr;
 use thoth_api::account::model::AccountDetails;
+use thoth_api::imprint::model::ImprintExtended as Imprint;
+use thoth_api::series::model::SeriesExtended as Series;
 use thoth_api::series::model::SeriesType;
 use uuid::Uuid;
 use yew::html;
@@ -27,7 +29,6 @@ use crate::models::imprint::imprints_query::FetchImprints;
 use crate::models::imprint::imprints_query::ImprintsRequest;
 use crate::models::imprint::imprints_query::ImprintsRequestBody;
 use crate::models::imprint::imprints_query::Variables as ImprintsVariables;
-use crate::models::imprint::Imprint;
 use crate::models::series::create_series_mutation::CreateSeriesRequest;
 use crate::models::series::create_series_mutation::CreateSeriesRequestBody;
 use crate::models::series::create_series_mutation::PushActionCreateSeries;
@@ -35,7 +36,6 @@ use crate::models::series::create_series_mutation::PushCreateSeries;
 use crate::models::series::create_series_mutation::Variables;
 use crate::models::series::series_types_query::FetchActionSeriesTypes;
 use crate::models::series::series_types_query::FetchSeriesTypes;
-use crate::models::series::Series;
 use crate::models::series::SeriesTypeValues;
 use crate::route::AdminRoute;
 use crate::route::AppRoute;
