+++ thoth-app/src/component/new_work.rs
@@ -1,5 +1,7 @@
 use std::str::FromStr;
 use thoth_api::account::model::AccountDetails;
+use thoth_api::imprint::model::ImprintExtended as Imprint;
+use thoth_api::work::model::WorkExtended as Work;
 use thoth_api::work::model::WorkStatus;
 use thoth_api::work::model::WorkType;
 use uuid::Uuid;
@@ -32,7 +34,6 @@ use crate::models::imprint::imprints_query::FetchImprints;
 use crate::models::imprint::imprints_query::ImprintsRequest;
 use crate::models::imprint::imprints_query::ImprintsRequestBody;
 use crate::models::imprint::imprints_query::Variables as ImprintsVariables;
-use crate::models::imprint::Imprint;
 use crate::models::work::create_work_mutation::CreateWorkRequest;
 use crate::models::work::create_work_mutation::CreateWorkRequestBody;
 use crate::models::work::create_work_mutation::PushActionCreateWork;
@@ -42,7 +43,6 @@ use crate::models::work::work_statuses_query::FetchActionWorkStatuses;
 use crate::models::work::work_statuses_query::FetchWorkStatuses;
 use crate::models::work::work_types_query::FetchActionWorkTypes;
 use crate::models::work::work_types_query::FetchWorkTypes;
-use crate::models::work::Work;
 use crate::models::work::WorkStatusValues;
 use crate::models::work::WorkTypeValues;
 use crate::route::AdminRoute;
