+++ thoth-app/src/component/series.rs
@@ -1,5 +1,7 @@
 use std::str::FromStr;
 use thoth_api::account::model::AccountDetails;
+use thoth_api::imprint::model::ImprintExtended as Imprint;
+use thoth_api::series::model::SeriesExtended as Series;
 use thoth_api::series::model::SeriesType;
 use uuid::Uuid;
 use yew::html;
@@ -29,7 +31,6 @@ use crate::models::imprint::imprints_query::FetchImprints;
 use crate::models::imprint::imprints_query::ImprintsRequest;
 use crate::models::imprint::imprints_query::ImprintsRequestBody;
 use crate::models::imprint::imprints_query::Variables as ImprintsVariables;
-use crate::models::imprint::Imprint;
 use crate::models::series::delete_series_mutation::DeleteSeriesRequest;
 use crate::models::series::delete_series_mutation::DeleteSeriesRequestBody;
 use crate::models::series::delete_series_mutation::PushActionDeleteSeries;
@@ -47,7 +48,6 @@ use crate::models::series::update_series_mutation::PushUpdateSeries;
 use crate::models::series::update_series_mutation::UpdateSeriesRequest;
 use crate::models::series::update_series_mutation::UpdateSeriesRequestBody;
 use crate::models::series::update_series_mutation::Variables as UpdateVariables;
-use crate::models::series::Series;
 use crate::models::series::SeriesTypeValues;
 use crate::route::AdminRoute;
 use crate::route::AppRoute;
