+++ thoth-app/src/component/work.rs
@@ -1,8 +1,13 @@
 use std::str::FromStr;
 use thoth_api::account::model::AccountDetails;
 use thoth_api::contribution::model::Contribution;
+use thoth_api::funding::model::FundingExtended as Funding;
+use thoth_api::imprint::model::ImprintExtended as Imprint;
+use thoth_api::issue::model::IssueExtended as Issue;
 use thoth_api::language::model::Language;
+use thoth_api::publication::model::PublicationExtended as Publication;
 use thoth_api::subject::model::Subject;
+use thoth_api::work::model::WorkExtended as Work;
 use thoth_api::work::model::WorkStatus;
 use thoth_api::work::model::WorkType;
 use uuid::Uuid;
@@ -38,10 +43,6 @@ use crate::component::utils::FormUrlInput;
 use crate::component::utils::FormWorkStatusSelect;
 use crate::component::utils::FormWorkTypeSelect;
 use crate::component::utils::Loader;
-use crate::models::funding::Funding;
-use crate::models::imprint::Imprint;
-use crate::models::issue::Issue;
-use crate::models::publication::Publication;
 use crate::models::work::delete_work_mutation::DeleteWorkRequest;
 use crate::models::work::delete_work_mutation::DeleteWorkRequestBody;
 use crate::models::work::delete_work_mutation::PushActionDeleteWork;
@@ -57,7 +58,6 @@ use crate::models::work::work_query::FetchWork;
 use crate::models::work::work_query::Variables;
 use crate::models::work::work_query::WorkRequest;
 use crate::models::work::work_query::WorkRequestBody;
-use crate::models::work::Work;
 use crate::models::work::WorkStatusValues;
 use crate::models::work::WorkTypeValues;
 use crate::route::AdminRoute;
