+++ thoth-app/src/models/imprint/mod.rs
@@ -1,9 +1,4 @@
-use chrono::DateTime;
-use chrono::Utc;
-use serde::Deserialize;
-use serde::Serialize;
-use thoth_api::publisher::model::Publisher;
-use uuid::Uuid;
+use thoth_api::imprint::model::ImprintExtended as Imprint;
 use yew::html;
 use yew::prelude::Html;
 use yew::Callback;
@@ -12,16 +7,6 @@ use yew::MouseEvent;
 use super::MetadataObject;
 use crate::route::AdminRoute;
 use crate::route::AppRoute;
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct Imprint {
-    pub imprint_id: Uuid,
-    pub imprint_name: String,
-    pub imprint_url: Option<String>,
-    pub updated_at: DateTime<Utc>,
-    pub publisher: Publisher,
-}
 
 impl MetadataObject for Imprint {
     fn create_route() -> AppRoute {
@@ -45,18 +30,6 @@ impl MetadataObject for Imprint {
                 <td>{imprint_url}</td>
                 <td>{&self.updated_at.format("%F %T")}</td>
             </tr>
-        }
-    }
-}
-
-impl Default for Imprint {
-    fn default() -> Imprint {
-        Imprint {
-            imprint_id: Default::default(),
-            imprint_name: "".to_string(),
-            imprint_url: None,
-            updated_at: chrono::TimeZone::timestamp(&Utc, 0, 0),
-            publisher: Default::default(),
         }
     }
 }
