+++ thoth-app/src/models/series/mod.rs
@@ -1,33 +1,16 @@
-use chrono::DateTime;
-use chrono::Utc;
 use serde::Deserialize;
 use serde::Serialize;
-use std::fmt;
+use thoth_api::series::model::SeriesExtended as Series;
 use thoth_api::series::model::SeriesType;
-use uuid::Uuid;
 use yew::html;
 use yew::prelude::Html;
 use yew::Callback;
 use yew::MouseEvent;
 
-use super::imprint::Imprint;
 use super::Dropdown;
 use super::MetadataObject;
 use crate::route::AdminRoute;
 use crate::route::AppRoute;
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct Series {
-    pub series_id: Uuid,
-    pub series_type: SeriesType,
-    pub series_name: String,
-    pub issn_print: String,
-    pub issn_digital: String,
-    pub series_url: Option<String>,
-    pub updated_at: DateTime<Utc>,
-    pub imprint: Imprint,
-}
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -39,31 +22,6 @@ pub struct SeriesTypeDefinition {
 #[serde(rename_all = "camelCase")]
 pub struct SeriesTypeValues {
     pub name: SeriesType,
-}
-
-impl Default for Series {
-    fn default() -> Series {
-        Series {
-            series_id: Default::default(),
-            series_type: Default::default(),
-            series_name: "".to_string(),
-            issn_print: "".to_string(),
-            issn_digital: "".to_string(),
-            series_url: None,
-            updated_at: chrono::TimeZone::timestamp(&Utc, 0, 0),
-            imprint: Default::default(),
-        }
-    }
-}
-
-impl fmt::Display for Series {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{} ({}, {})",
-            self.series_name, self.issn_print, self.issn_digital
-        )
-    }
 }
 
 impl Dropdown for Series {}
