+++ thoth-app/src/models/series/serieses_query.rs
@@ -1,8 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
+use thoth_api::series::model::SeriesExtended as Series;
 use thoth_api::series::model::SeriesOrderBy;
-
-use super::Series;
 
 pub const SERIESES_QUERY: &str = "
     query SeriesesQuery($limit: Int, $offset: Int, $filter: String, $publishers: [Uuid!], $order: SeriesOrderBy) {
